Remove tabs from the search page
The purpose of a separate list page was unclear and no work on it
had yet been done, so it was decided that it should be removed.

This commit also removes the tab components that were defined as
part of this repository; in any future cases where we need such
a component, the ones built into HDS should be used instead.