Remove invalid margin style rule
rem-calc is only used in the other UI, and it seems to be parsed
inconsistently from the SCSS files - in the dev environment, the rule
was added to the transpiled file as is (and ignored by browsers),
but it was converted to the given pixel value instead in the
'production' build on the staging server. The rule is, regardless,
redundant.