Fix issue of fetching favourite while updating token
Observed behavior previously was that when token is to be refreshed
the useEffect  is triggered when tokenOutdated. The same useEffect
also happens to always run fetchFavourite whenever favourite is not being
fetched.
This lead to fetchFavourite being run unsuccesfully on every token refresh,
while fetchFavourite used the old token and never succeeded.