Use only oidc user for setting user logged in
AuthdependentContext asummed that loggedIn means having oidc user, and api token.
Changed this to only take into account oidc user, because api token
is not part of the oauth login flow, it is fetched separately.
Having access token allows fetching api token, and as long as it exist
and is valid, it should be possible to renew the api token.
So in short it is assumed now that logged in means having the
oidc user.